Refactored the log in link/dialog into separate files for better maintainability
