+++ client/src/App.js
@@ -30,7 +30,6 @@ const client = new ApolloClient({
 });
 
 function App() {
-  console.log(document.location);
   const windowSize = useWindowSize();
 
   return (
