+++ client/src/components/Header.js
@@ -1,7 +1,7 @@
 import { Link } from 'react-router-dom';
 import React from 'react';
 import './Header.css';
-import LogIn from './LogIn';
+import LogInDialog from './LogIn';
 import LogOut from './LogOut';
 
 import Auth from '../utils/auth';
@@ -17,7 +17,7 @@ const Header = () => {
         <h1>Will It Fit</h1>
       </Link>
 
-      {Auth.loggedIn() ? <LogOut /> : <LogIn />}
+      {Auth.loggedIn() ? <LogOut /> : <LogInDialog />}
     </header>
   );
 };
